WoT restructure A2: sizing/ -> axisSizing/ (git-mv + imports)
