+++ handsontable/src/3rdparty/walkontable/src/axisSizing/axisSizeSource.ts
