+++ handsontable/src/3rdparty/walkontable/src/axisSizing/boxModel.ts
