+++ handsontable/src/3rdparty/walkontable/src/axisSizing/defaultSizeSource.ts
