+++ handsontable/src/3rdparty/walkontable/src/utils/column.ts
@@ -1,5 +1,5 @@
 import type { TableDeps } from '../table';
-import type { ColumnSizeSource } from '../sizing/axisSizeSource';
+import type { ColumnSizeSource } from '../axisSizing/axisSizeSource';
 /**
  * Column utils class contains all necessary information about sizes of the columns.
  *
