+++ handsontable/src/3rdparty/walkontable/src/utils/row.ts
@@ -1,5 +1,5 @@
 import type { TableDeps } from '../table';
-import type { RowSizeSource } from '../sizing/axisSizeSource';
+import type { RowSizeSource } from '../axisSizing/axisSizeSource';
 /**
  * Row utils class contains all necessary information about sizes of the rows.
  *
